01-auth-profile-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.32
psycopg2-binary==2.9.9

pydantic-settings==2.4.0
pydantic[email]==2.10.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
